+++ db/migrate/20160628174954_remove_status_from_subject_events.rb
@@ -1,0 +1,9 @@
+class RemoveStatusFromSubjectEvents < ActiveRecord::Migration
+  def up
+    remove_column :subject_events, :status
+  end
+
+  def down
+    add_column :subject_events, :status, :string, null: false, default: 'scheduled'
+  end
+end